Use autounboxing for Cassandra target ports

+++ instrumentation/cassandra/cassandra-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/cassandra/v3_0/CassandraConfiguredTarget.java
@@ -69,8 +69,7 @@ public class CassandraConfiguredTarget {
 
     DbServerTargetBuilder targetBuilder = DbServerTarget.builder(DEFAULT_PORT);
     for (ContactPoint point : contactPoints.points) {
-      targetBuilder.addEndpoint(
-          point.host, point.port == null ? configuredPort : point.port.intValue());
+      targetBuilder.addEndpoint(point.host, point.port == null ? configuredPort : point.port);
     }
     return targetBuilder.build();
   }
